(PC-18725) feat(onboarding): Move CreditStatus enum to types file
