+++ src/features/onboarding/types.ts
@@ -1,0 +1,5 @@
+export enum CreditStatus {
+  GONE = 'passé',
+  ONGOING = 'cette année',
+  COMING = 'à venir',
+}
